# src/main/java/ru/easybot/testTusk/models/entities/Display.java
package ru.easybot.testTusk.models.entities;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import lombok.*;
import org.hibernate.annotations.BatchSize;

import javax.persistence.Entity;
@Entity
@Getter
@Setter
@ToString
@NoArgsConstructor
@AllArgsConstructor
@JsonIgnoreProperties(ignoreUnknown = true)
@BatchSize(size = 10)
public class Display extends Product{
    private Long diagonal;
}