# src/main/java/ru/easybot/testTusk/models/entities/HardDisk.java
package ru.easybot.testTusk.models.entities;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
import lombok.*;
import org.hibernate.annotations.BatchSize;

import javax.persistence.Entity;

@Entity
@Getter
@Setter
@ToString
@NoArgsConstructor
@AllArgsConstructor
@JsonIgnoreProperties(ignoreUnknown = true)
@BatchSize(size = 10)
public class HardDisk extends Product {
    private Long capacity;
}